+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/ast/ApexQualifiedName.java
@@ -9,8 +9,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
-
 
 /**
  * Qualified name of an apex class or method.
@@ -19,14 +17,11 @@ import org.apache.commons.lang3.StringUtils;
  */
 public final class ApexQualifiedName {
 
-
-    private final String nameSpace;
     private final String[] classes;
     private final String operation;
 
 
-    private ApexQualifiedName(String nameSpace, String[] classes, String operation) {
-        this.nameSpace = nameSpace;
+    private ApexQualifiedName(String[] classes, String operation) {
         this.operation = operation;
         this.classes = classes;
     }
@@ -39,16 +34,6 @@ public final class ApexQualifiedName {
 
     public String[] getClasses() {
         return Arrays.copyOf(classes, classes.length);
-    }
-
-
-    /**
-     * Gets the namespace prefix of this resource.
-     *
-     * @return The namespace prefix
-     */
-    public String getNameSpace() {
-        return nameSpace;
     }
 
 
@@ -75,7 +60,6 @@ public final class ApexQualifiedName {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(nameSpace).append("__");
         sb.append(classes[0]);
 
         if (classes.length > 1) {
@@ -101,14 +85,13 @@ public final class ApexQualifiedName {
             return this;
         }
 
-        return new ApexQualifiedName(this.nameSpace, this.classes, null);
+        return new ApexQualifiedName(this.classes, null);
     }
 
 
     @Override
     public int hashCode() {
-        int result = nameSpace.hashCode();
-        result = 31 * result + Arrays.hashCode(classes);
+        int result = Arrays.hashCode(classes);
         result = 31 * result + (operation != null ? operation.hashCode() : 0);
         return result;
     }
@@ -118,8 +101,7 @@ public final class ApexQualifiedName {
     public boolean equals(Object obj) {
         return obj instanceof ApexQualifiedName
                && Objects.deepEquals(classes, ((ApexQualifiedName) obj).classes)
-               && Objects.equals(operation, ((ApexQualifiedName) obj).operation)
-               && Objects.equals(nameSpace, ((ApexQualifiedName) obj).nameSpace);
+               && Objects.equals(operation, ((ApexQualifiedName) obj).operation);
 
     }
 
@@ -129,24 +111,23 @@ public final class ApexQualifiedName {
      *
      * <p>Here are some examples of the format:
      * <ul>
-     * <li> {@code namespace__OuterClass.InnerClass}: name of an inner class
-     * <li> {@code namespace__Class#method(String, int)}: name of an operation
+     * <li> {@code OuterClass.InnerClass}: name of an inner class
+     * <li> {@code Class#method(String, int)}: name of an operation
      * </ul>
      *
      * @param toParse The string to parse
      *
      * @return An ApexQualifiedName, or null if the string couldn't be parsed
      */
-    // private static final Pattern FORMAT = Pattern.compile("(\\w+)__(\\w+)(.(\\w+))?(#(\\w+))?"); // TODO
+    // private static final Pattern FORMAT = Pattern.compile("(\\w+)(.(\\w+))?(#(\\w+))?"); // TODO
     public static ApexQualifiedName ofString(String toParse) {
         throw new UnsupportedOperationException();
     }
 
 
     static ApexQualifiedName ofOuterClass(ASTUserClassOrInterface<?> astUserClass) {
-        String ns = astUserClass.getNamespace();
         String[] classes = {astUserClass.getSimpleName()};
-        return new ApexQualifiedName(StringUtils.isEmpty(ns) ? "c" : ns, classes, null);
+        return new ApexQualifiedName(classes, null);
     }
 
 
@@ -154,21 +135,20 @@ public final class ApexQualifiedName {
 
         String[] classes = Arrays.copyOf(parent.classes, parent.classes.length + 1);
         classes[classes.length - 1] = astUserClass.getSimpleName();
-        return new ApexQualifiedName(parent.nameSpace, classes, null);
+        return new ApexQualifiedName(classes, null);
     }
 
 
     static ApexQualifiedName ofOuterEnum(ASTUserEnum astUserEnum) {
-        String ns = astUserEnum.getNamespace();
         String[] classes = {astUserEnum.getSimpleName()};
-        return new ApexQualifiedName(StringUtils.isEmpty(ns) ? "c" : ns, classes, null);
+        return new ApexQualifiedName(classes, null);
     }
 
 
     static ApexQualifiedName ofNestedEnum(ApexQualifiedName parent, ASTUserEnum astUserEnum) {
         String[] classes = Arrays.copyOf(parent.classes, parent.classes.length + 1);
         classes[classes.length - 1] = astUserEnum.getSimpleName();
-        return new ApexQualifiedName(parent.nameSpace, classes, null);
+        return new ApexQualifiedName(classes, null);
     }
 
     /**
@@ -204,21 +184,20 @@ public final class ApexQualifiedName {
         if (enumParent != null) {
             ApexQualifiedName baseName = enumParent.getQualifiedName();
 
-            return new ApexQualifiedName(baseName.nameSpace, baseName.classes, getOperationString(node));
+            return new ApexQualifiedName(baseName.classes, getOperationString(node));
         }
 
         ASTUserClassOrInterface<?> parent = node.ancestors(ASTUserClassOrInterface.class).firstOrThrow();
         if (parent instanceof ASTUserTrigger) {
             ASTUserTrigger trigger = (ASTUserTrigger) parent;
-            String ns = trigger.getNamespace();
             String targetObj = trigger.getTargetName();
 
-            return new ApexQualifiedName(StringUtils.isEmpty(ns) ? "c" : ns, new String[]{"trigger", targetObj}, trigger.getSimpleName()); // uses a reserved word as a class name to prevent clashes
+            return new ApexQualifiedName(new String[]{"trigger", targetObj}, trigger.getSimpleName()); // uses a reserved word as a class name to prevent clashes
 
         } else {
             ApexQualifiedName baseName = parent.getQualifiedName();
 
-            return new ApexQualifiedName(baseName.nameSpace, baseName.classes, getOperationString(node));
+            return new ApexQualifiedName(baseName.classes, getOperationString(node));
         }
     }
 }